[20] Documentar los controladores mediante Swagger

+++ src/main/java/es/unizar/tmdad/service/impl/RoomServiceImpl.java
@@ -4,6 +4,9 @@ import es.unizar.tmdad.adt.EventType;
 import es.unizar.tmdad.adt.UserEvent;
 import es.unizar.tmdad.dto.RoomCreationDto;
 import es.unizar.tmdad.dto.RoomDto;
+import es.unizar.tmdad.exception.EntityNotFoundException;
+import es.unizar.tmdad.exception.UnauthorizedException;
+import es.unizar.tmdad.exception.UserNotTheOwnerException;
 import es.unizar.tmdad.mapper.RoomMapper;
 import es.unizar.tmdad.repository.RoomRepository;
 import es.unizar.tmdad.repository.entity.RoomEntity;
@@ -52,11 +55,15 @@ public class RoomServiceImpl implements RoomService {
     public RoomDto getRoom(Long id) {
         return repository.findById(id)
                 .map(mapper::mapRoom)
-                .orElseThrow(() -> new RuntimeException("NOT FOUND: Error getting room with id " + id + "."));
+                .orElseThrow(() -> new EntityNotFoundException(EntityNotFoundException.EntityType.ROOM, "NOT FOUND: Error getting room with id " + id + "."));
     }
 
     @Override
-    public void deleteRoom(Long id) {
+    public void deleteRoom(Long id, String owner) throws UnauthorizedException {
+        var room = repository.findById(id).orElseThrow(() -> new EntityNotFoundException(EntityNotFoundException.EntityType.ROOM, String.valueOf(id)));
+        if(!Objects.equals(owner, room.getOwner().getEmail())){
+            throw new UnauthorizedException("User ("+owner+") is not the owner of the room, so he cannot delete it.");
+        }
         repository.deleteById(id);
 
         UserEvent event = UserEvent.builder()
@@ -74,7 +81,7 @@ public class RoomServiceImpl implements RoomService {
         UserEntity user = this.userService.getUserEntity(userId).orElseThrow();
 
         if(!Objects.equals(room.getOwner(), owner)){
-            throw new RuntimeException("Cannot add user to room if not the owner.");
+            throw new UserNotTheOwnerException("Cannot add user to room if not the owner.");
         }
 
         switch (op){
